Move game helper functions into util.js
