+++ src/index.js
@@ -1,8 +1,10 @@
-const { Variant, Player, Board, reverseMove, rollDie } = require("./game");
+const { Variant, Player, random, reverseMove, rollDie } = require("./util");
+const { Board } = require("./game");
 
 exports.Variant = Variant;
 exports.Player = Player;
 exports.Board = Board;
+exports.random = random;
 exports.reverseMove = reverseMove;
 exports.rollDie = rollDie;
